Moved classes Image and Pixel to separate file image.py
